requirements.txt
Flask
gunicorn
flask-jwt-extended
mysql-connector-python
python-dotenv
flask-cors
werkzeug
